Split images.py into multiple files
